Move ledger UI functions into UI
This makes the ledger module platform independent
